Split Screenshot god classes under 500 lines
Extract Class on ScreenshotUseCase, UIElementAnnotator, and RaycastGridAnnotator so capture orchestration, annotation rendering/styling, and raycast summary/physics builders each have one clear responsibility without changing screenshot behavior.
