+++ Packages/src/Editor/FirstPartyTools/Screenshot/ScreenshotFileWriter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b54dc29221946e4af50763527638e22
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
