+++ Packages/src/Editor/FirstPartyTools/Screenshot/ScreenshotParameterValidator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1589cc00a58e4863af06b2d5d690d824
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
