+++ Packages/src/Editor/FirstPartyTools/Screenshot/UIElementAnnotationStyling.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9619dc22a1374af4af5fbd787d7be66e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
